# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for pollution-tracker app",
        "title": "pollution-tracker API",
        "contact": {}
    },
    "paths": {
        "/api/anomalies": {
            "get": {
                "description": "Gets anomalies for a given time range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "anomalies"
                ],
                "summary": "Gets anomalies for range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start time",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End time",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Anomalies",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/pollution.Pollution"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid params",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch pollution entries from database",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/pollutants": {
            "get": {
                "description": "Gets distinct pollutants that exists in database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pollutants"
                ],
                "summary": "Gets pollutants",
                "responses": {
                    "200": {
                        "description": "Pollutants",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch pollutants entries from database",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/pollutions": {
            "get": {
                "description": "Gets all pollution values for given time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pollutions"
                ],
                "summary": "Gets pollution values",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start time",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End time",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pollution values",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/pollution.Pollution"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid params",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch pollution entries from database",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Posts a new pollution entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pollutions"
                ],
                "summary": "Posts pollution entry",
                "parameters": [
                    {
                        "description": "Request of adding a new pollution entry",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pollution.Pollution"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully received the pollution entry",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Failed to marshal request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to publish pollution entry to RabbitMQ queue",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/pollutions/density/rect": {
            "get": {
                "description": "Gets pollution densities for a given rect and time range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pollutions"
                ],
                "summary": "Gets pollution densities of rect",
                "parameters": [
                    {
                        "type": "number",
                        "description": "latFrom",
                        "name": "latFrom",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "latTo",
                        "name": "latTo",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "longFrom",
                        "name": "longFrom",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "longTo",
                        "name": "longTo",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "from",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "to",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pollutant",
                        "name": "pollutant",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pollution Densities",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/pollution.PollutionDensity"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid params",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch pollution entries from database",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/pollutions/{latitude}/{longitude}": {
            "get": {
                "description": "Gets pollution values for given location and time range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pollutions"
                ],
                "summary": "Gets pollution values",
                "parameters": [
                    {
                        "type": "string",
                        "description": "latitude",
                        "name": "latitude",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "longitude",
                        "name": "longitude",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pollution Values",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/pollution.PollutionValueResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid params",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch pollution entries from database",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "pollution.Pollution": {
            "type": "object",
            "properties": {
                "is_anomaly": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "pollutant": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "pollution.PollutionDensity": {
            "type": "object",
            "properties": {
                "density": {
                    "type": "number"
                },
                "pollutant": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "pollution.PollutionValueResponse": {
            "type": "object",
            "properties": {
                "pollutant": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}